+++ build-support/globalize-orbit-jsonapi.js
@@ -1,0 +1,1 @@
+window.OrbitJSONAPI = requireModule("orbit-jsonapi");
