+++ src/orbit-jsonapi.js
@@ -1,0 +1,3 @@
+import JSONAPISource from 'orbit-jsonapi/jsonapi-source';
+
+export { JSONAPISource };
