+++ test/tests/integration/pessimistic-sync-test.js
@@ -1,6 +1,6 @@
 import Orbit from 'orbit/main';
 import Store from 'orbit-common/store';
-import JsonApiSource from 'orbit-common/jsonapi-source';
+import JSONAPISource from 'orbit-jsonapi/jsonapi-source';
 import LocalStorageSource from 'orbit-local-storage/local-storage-source';
 import qb from 'orbit-common/query/builder';
 import KeyMap from 'orbit-common/key-map';
@@ -33,7 +33,7 @@ module('Integration - Pessimistic Sync', function(hooks) {
     fetchStub = sinon.stub(Orbit, 'fetch');
 
     let keyMap = new KeyMap();
-    remote = new JsonApiSource({ schema: planetsSchema, keyMap: new KeyMap() });
+    remote = new JSONAPISource({ schema: planetsSchema, keyMap: new KeyMap() });
     store = new Store({ schema: planetsSchema, keyMap });
     backup = new LocalStorageSource({ schema: planetsSchema, keyMap });
